+++ src/scripts/_3_NPS.py
@@ -11,31 +11,30 @@
 import arcpy
 from ._1b_add_fields import AddFields
 from ._2b_assign_domains import AssignDomains
+# add 2j standardize domains, 2f Categories
 from ._7a_enrichments_polygon import enrich_polygons
 from ._2k_keep_fields import KeepFields
-# from sys import argv
-from .utils import init_gdb, delete_scratch_files, runner
-# import os
+from .utils import init_gdb, delete_scratch_files
+import os
 import time
 
-original_gdb, workspace, scratch_workspace = init_gdb()
-
+workspace, scratch_workspace = init_gdb()
+# TODO add print steps, rename variables
 
 # def NPS(nps_flat_fuels_enriched_20221102="C:\\Users\\sageg\\Documents\\ArcGIS\\Projects\\PC414 CWI Million Acres\\PC414 CWI Million Acres.gdb\\d_Enriched\\nps_flat_fuels_enriched_20221102", usfs_haz_fuels_treatments_standardized_20220713_2_="C:\\Users\\sageg\\Documents\\ArcGIS\\Projects\\PC414 CWI Million Acres\\PC414 CWI Million Acres.gdb\\c_Standardized\\nps_flat_fuels_standardized_20221102"):  # 6r NPS 20221123
-def NPS(input_fc, output_standardized, output_enriched):
+def NPS(input_fc, output_standardized, output_enriched, delete_scratch=True):
     start = time.time()
     print(f"Start Time {time.ctime()}")
     with arcpy.EnvManager(
+        workspace=workspace,
+        scratchWorkspace=scratch_workspace, 
         outputCoordinateSystem= arcpy.SpatialReference("NAD 1983 California (Teale) Albers (Meters)"), #WKID 3310
         cartographicCoordinateSystem=arcpy.SpatialReference("NAD 1983 California (Teale) Albers (Meters)"), #WKID 3310
-        extent="""450000, -374900, 540100, -604500,
-                  DATUM["NAD 1983 California (Teale) Albers (Meters)"]""",
+        extent="xmin=-374900, ymin=-604500, xmax=540100, ymax=450000, spatial_reference='NAD 1983 California (Teale) Albers (Meters)'", 
         preserveGlobalIds=True, 
         qualifiedFieldNames=False, 
-        scratchWorkspace=scratch_workspace, 
         transferDomains=False, 
-        transferGDBAttributeProperties=True, 
-        workspace=workspace,
+        transferGDBAttributeProperties=False, 
         overwriteOutput = True,
     ):
 
@@ -125,7 +124,7 @@ def NPS(input_fc, output_standardized, output_enriched):
             clear_field_alias="DO_NOT_CLEAR",
         )
 
-        # Process: 1b Add Fields (1b Add Fields) (PC414CWIMillionAcres)
+        # Process: 1b Add Fields (1b Add Fields)
         nps_addfields = AddFields(Input_Table=nps_altered_prjid)
 
         # Process: Calculate Project ID (Calculate Field) (management)
@@ -423,7 +422,7 @@ def NPS(input_fc, output_standardized, output_enriched):
         nps_standardized_keepfields = KeepFields(output_standardized)
 
         print("Performing Enrichments")
-        # Process: 7a Enrichments Polygon (2) (7a Enrichments Polygon) (PC414CWIMillionAcres)
+        # Process: 7a Enrichments Polygon (2) (7a Enrichments Polygon)
         enrich_polygons(
             enrich_out=nps_enriched_scratch, enrich_in=nps_standardized_keepfields
         )
@@ -451,28 +450,13 @@ def NPS(input_fc, output_standardized, output_enriched):
             enforce_domains="NO_ENFORCE_DOMAINS",
         )
 
-        # Process: 2b Assign Domains (2b Assign Domains) (PC414CWIMillionAcres)
+        # Process: 2b Assign Domains (2b Assign Domains)
         # nps_enriched_assign_domains =
         AssignDomains(in_table=output_enriched)
 
-        print("Deleting Scratch Files")
-        delete_scratch_files(
-            gdb=scratch_workspace, delete_fc="yes", delete_table="yes", delete_ds="yes"
-        )
+        if delete_scratch: delete_scratch_files(
+                gdb=scratch_workspace, delete_fc="yes", delete_table="yes", delete_ds="yes"
+            )
 
         end = time.time()
         print(f"Time Elapsed: {(end-start)/60} minutes")
-
-
-# if __name__ == "__main__":
-#     runner(workspace, scratch_workspace, NPS, "*argv[1:]")
-    # # Global Environment settings
-    # with arcpy.EnvManager(
-    # extent="""-124.415162172178 32.5342699477235 -114.131212866967 42.0095193288898 GEOGCS["GCS_WGS_1984",DATUM["D_WGS_1984",SPHEROID["WGS_1984",6378137.0,298.257223563]],PRIMEM["Greenwich",0.0],UNIT["Degree",0.0174532925199433]]""",  outputCoordinateSystem="""PROJCS["NAD_1983_California_Teale_Albers",GEOGCS["GCS_North_American_1983",DATUM["D_North_American_1983",SPHEROID["GRS_1980",6378137.0,298.257222101]],PRIMEM["Greenwich",0.0],UNIT["Degree",0.0174532925199433]],PROJECTION["Albers"],PARAMETER["False_Easting",0.0],PARAMETER["False_Northing",-4000000.0],PARAMETER["Central_Meridian",-120.0],PARAMETER["Standard_Parallel_1",34.0],PARAMETER["Standard_Parallel_2",40.5],PARAMETER["Latitude_Of_Origin",0.0],UNIT["Meter",1.0]]""",
-    # preserveGlobalIds=True,
-    # qualifiedFieldNames=False,
-    # scratchWorkspace=scratch_workspace,
-    # transferDomains=True,
-    # transferGDBAttributeProperties=True,
-    # workspace=workspace):
-    #     NPS(*argv[1:])
